requirements.txt
Django>=5.0
djangorestframework>=3.14
openai>=1.0
python-dotenv>=1.0
scikit-learn>=1.3
numpy>=1.24
joblib>=1.3
django-crispy-forms>=2.1
crispy-bootstrap5>=2024.2
dj-database-url>=2.1
gunicorn>=21.2
whitenoise>=6.6
reportlab>=4.0
Pillow>=10.0
